# proof of concept: allow using Pardiso as a linear solver