Pardiso is a sparse linear solver that comes with MKL and can have a significant speed improvement over the default solver in julia (UMFPACK).

This PR is a proof of concept of using Pardiso in the stacked time solver.
Running the FRB US model with the noise it performs around 2x better on my machine compared to the default:

```
Section            ncalls     time    %tot     avg     alloc    %tot      avg
─────────────────────────────────────────────────────────────────────────────
factorize           8    3.66s   57.9%   458ms   34.7MiB    1.8%  4.34MiB

StateSpaceEcon.StackedTimeSolver.USE_PARDISO_FOR_LU = false:

Section            ncalls     time    %tot     avg     alloc    %tot      avg
─────────────────────────────────────────────────────────────────────────────
factorize           8    8.43s   82.3%   1.05s   5.01GiB   73.4%   642MiB
```