@bbejanov 2023-03-09
Thanks! It's nice to have this option and we'll definitely include it in StateSpaceEcon. 

For our internal use, we'll probably need a commercial license for Pardiso - I don't think it's free and open source. We'll ask our legal services to check that.