@KristofferC 2023-03-10
I want to write a little bit more here about further gains that can be had with Pardiso.
I haven't implemented it yet because I still don't have the full picture in how the Jacobian changes during the simulation.

Pardiso has three stages:

1. Symbolical factorization (only requires the sparsity pattern of the matrix, the actual values in the matrix are unused).
2. Numerical factorization (uses the numerical values in the matrix)
3. Solving (uses the RHS)

From my timings (see below) step 1 takes about 75%, step 2 takes about 25% and step 3 is fast.

So the first step to be efficient is to only do the numerical factorization when the numerical values in the matrix have been updated.
I think the current code already does this.

But! Even if the numerical values in the matrix have changed, if the sparsity pattern is the same, that is, the same indices in the matrix is
non-zero, the symbolical factorization is still valid and only step 2 and 3 has to be redone. Since step 1 is the most expensive one this can 
have a quite dramatic speedup. It is quite common for the sparsity pattern to be constant along a simulation since the "structure" of the problem
tend not to change.

----------


I stored the matrix in Matrix Market format (https://github.com/JuliaSparse/MatrixMarket.jl) and attached it to this post as a zip file ([J.zip](https://github.com/bankofcanada/StateSpaceEcon.jl/files/10939884/J.zip)) (required by github).
The code below sets up some functions that times the different steps in the solution process, loads the matrix and times it.

```julia
mutable struct PardisoFactorization
    ps::MKLPardisoSolver
    J::SparseMatrixCSC
end

# See https://github.com/JuliaSparse/Pardiso.jl/blob/master/examples/exampleunsym.jl
function pardiso_factorize(JJ)
    ps = MKLPardisoSolver()
    set_matrixtype!(ps, Pardiso.REAL_NONSYM)
    pardisoinit(ps)
    fix_iparm!(ps, :N)
    # See https://www.intel.com/content/www/us/en/develop/documentation/onemkl-developer-reference-c/top/sparse-solver-routines/onemkl-pardiso-parallel-direct-sparse-solver-iface/pardiso-iparm-parameter.html
    set_iparm!(ps, 2, 2) # The parallel (OpenMP) version of the nested dissection algorithm. 
    JJ_pardiso = get_matrix(ps, JJ, :N)

    @info "Step 1: Analysis"
    set_phase!(ps, Pardiso.ANALYSIS)
    @time pardiso(ps, JJ_pardiso, Float64[])

    pf = PardisoFactorization(ps, JJ_pardiso)
    finalizer(pf) do x
        set_phase!(x.ps, Pardiso.RELEASE_ALL)
        pardiso(x.ps)
    end
end

function pardiso_solve!(ps::PardisoFactorization, x::AbstractVector)
    ps, J = ps.ps, ps.J

    @info "Step 2: Numerical factorization:"
    set_phase!(ps, Pardiso.NUM_FACT)
    @time pardiso(ps, J, Float64[])


    @info "Step 3: Numerical factorization:"
    set_phase!(ps, Pardiso.SOLVE_ITERATIVE_REFINE)
    X = similar(x) # Solution is stored in X
    @time pardiso(ps, X, J, x)
    copy!(x, X)
end
```

```julia
julia> using MatrixMarket

julia> J = MatrixMarket.mmread("J.mtx")

julia> ps = pardiso_factorize(J);
[ Info: Step 1: Analysis
  0.460367 seconds (10 allocations: 4.340 MiB)

julia> pardiso_solve!(ps, copy(b));
[ Info: Step 2: Numerical factorization:
  0.123646 seconds (10 allocations: 4.340 MiB, 2.20% gc time)
[ Info: Step 3: Numerical factorization:
  0.008820 seconds (9 allocations: 4.703 MiB)
```