@KristofferC 2023-03-10
I implemented the reusing of the symbolic factorization in the last commit. It is a bit ugly and could use some cleanup but the I mostly wanted to see the performance change and it is significant:

Before the reuse  of the factorization I get:

```
 ─────────────────────────────────────────────────────────────────────────────
                                     Time                    Allocations      
                            ───────────────────────   ────────────────────────
      Tot / % measured:          5.86s /  99.3%           0.95GiB /  89.6%    

 Section            ncalls     time    %tot     avg     alloc    %tot      avg
 ─────────────────────────────────────────────────────────────────────────────
   stackedtime_RJ        8    5.74s   98.6%   718ms    810MiB   93.3%   101MiB
     factorize           8    4.70s   80.7%   587ms   69.5MiB    8.0%  8.68MiB
     eval_RJ         1.90k    995ms   17.1%   525μs    664MiB   76.5%   358KiB
 ─────────────────────────────────────────────────────────────────────────────
```

After the reuse I get:

```
 ───────────────────────────────────────────────────────────────────────────
                                   Time                    Allocations      
                          ───────────────────────   ────────────────────────
     Tot / % measured:         2.71s /  96.9%            936MiB /  88.2%    

 Section          ncalls     time    %tot     avg     alloc    %tot      avg
 ───────────────────────────────────────────────────────────────────────────
 stackedtime_RJ        8    2.54s   96.6%   317ms    779MiB   94.4%  97.4MiB
   factorize           8    1.48s   56.2%   185ms   39.1MiB    4.7%  4.88MiB
   eval_RJ         1.90k    1.02s   38.9%   539μs    664MiB   80.4%   358KiB
 ───────────────────────────────────────────────────────────────────────────
```