@bbejanov 2023-04-05
Thanks @KristofferC, this is a great idea.

I figured out that it is possible to cache the symbolic step and reuse it with UMFPACK as well.  So I'll add this and I'll also keep your version with pardiso and integrate the two so users can easily switch between umfpack and pardiso. I'll do that in a separate PR.  Thanks again!